Add auth checks

+++ packages/api/src/controllers/site/delete.js
@@ -1,8 +1,13 @@
 import { withoutInternalKeys } from '@layouthq/util'
+import { authorize, MANAGE_SITE } from '../../auth'
 
-export default async ({ db, params }, res) => {
+export default async ({ db, params, authInfo }, res) => {
   try {
     const { id } = params
+
+    if (!authorize(authInfo, [MANAGE_SITE], [Number(id)])) {
+      return res.status(403).send(JSON.stringify({ message: `You are not authorized to delete site id ${id}.` }))
+    }
 
     const sites = await db.collection('sites')
     const { value: site } = await sites.findOneAndDelete({ id: Number(id) })
